% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{agreement}
\alias{agreement}
\title{Agreement of two vectors at a given depth}
\usage{
agreement(x, y, depth = min(length(x), length(y)))
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{depth}{A number.}
}
\value{
The agreement of \code{x} and \code{y} to \code{depth}.
}
\description{
Agreement of two vectors at a given depth
}

