% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{intersection}
\alias{intersection}
\title{Intersection of two vectors at a given depth}
\usage{
intersection(x, y, depth = min(length(x), length(y)))
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{depth}{A number.}
}
\value{
The intersection of \code{x} and \code{y} to \code{depth}.
}
\description{
Intersection of two vectors at a given depth
}

