% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{overlap}
\alias{overlap}
\title{Overlap of two vectors at a given depth}
\usage{
overlap(x, y, depth = min(length(x), length(y)))
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{depth}{A number.}
}
\value{
The overlap of \code{x} and \code{y} to \code{depth}.
}
\description{
Overlap of two vectors at a given depth
}

