% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbo.R
\name{rbo_ext}
\alias{rbo_ext}
\title{Extrapolated Rank-Biased Overlap}
\usage{
rbo_ext(x, y, p = 0.9)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{p}{A number, representing persistence.}
}
\description{
Extrapolated Rank-Biased Overlap
}

