% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbo.R
\name{rbo_ext_uneven}
\alias{rbo_ext_uneven}
\title{Extrapolated Rank-Biased Overlap on uneven-length vectors}
\usage{
rbo_ext_uneven(l, s, p = 0.9)
}
\arguments{
\item{l}{A vector.}

\item{s}{A vector, shorter than \code{l}.}

\item{p}{A number, representing persistence.}
}
\description{
Extrapolated Rank-Biased Overlap on uneven-length vectors
}

